+++ both/lib/FM.jsx
@@ -1,6 +1,6 @@
 Schema = {};
 FM = {
-	version:"1.3.4b",
+	version:"1.3.5b",
 	collections:{},
 	schemas:{}
 }